+++ integration-tests/pkg/collector/collector_docker.go
@@ -88,20 +88,21 @@ func (c *DockerCollectorManager) TearDown() error {
 
 	if !isRunning {
 		logs, logsErr := c.captureLogs("collector")
-		logsEnd := logs
-		if logsErr == nil {
-			logsSplit := strings.Split(logs, "\n")
-			logsEnd = strings.Join(logsSplit[max(0, len(logsSplit)-24):], "\n")
-			logsEnd = fmt.Sprintf("\ncollector logs:\n%s\n", logsEnd)
-		}
+
 		// Check if collector container segfaulted or exited with error
 		exitCode, err := c.executor.ExitCode(executor.ContainerFilter{
 			Name: "collector",
 		})
-		if err != nil {
-			return fmt.Errorf("Failed to get container exit code%s: %w", logsEnd, err)
-		}
-		if exitCode != 0 {
+		if err != nil || exitCode != 0 {
+			logsEnd := ""
+			if logsErr == nil {
+				logsSplit := strings.Split(logs, "\n")
+				logsEnd = fmt.Sprintf("\ncollector logs:\n%s\n",
+					strings.Join(logsSplit[max(0, len(logsSplit)-24):], "\n"))
+			}
+			if err != nil {
+				return fmt.Errorf("Failed to get container exit code%s: %w", logsEnd, err)
+			}
 			return fmt.Errorf("Collector container has non-zero exit code (%d)%s", exitCode, logsEnd)
 		}
 	} else {
